% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plm_labels}
\alias{plm_labels}
\title{Public land lanbels}
\format{
A vector of labels from the open-data-platform:plm25 dataset
}
\usage{
data(plm_labels)
}
\description{
Public land management data labels
}
\details{
\itemize{
\item plm_labels. labels from the open-data-platform:plm25 dataset
}
}
\keyword{datasets}
