% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_simulation.R
\name{precision_simulation}
\alias{precision_simulation}
\title{Generate a simulation of abundance estimates for selected sites}
\usage{
precision_simulation(
  deployment_weeks = 10,
  sampling_area,
  survey_area,
  n_sites,
  species = c("Sambar", "Fallow", "Red", "Hog"),
  sampling_type = "hexagonal"
)
}
\arguments{
\item{deployment_weeks}{weeks to deploy cameras for}

\item{sampling_area}{area to sample within (i.e., can be resricted around roads)}

\item{survey_area}{total area to estimate abundance for (e.g. full National Park)}

\item{n_sites}{number of camera sites}

\item{species}{species (Sambar, Fallow, Red and Hog)}

\item{sampling_type}{type of sampling to use (hexagonal or random)}
}
\value{
list
}
\description{
Generate a simulation of abundance estimates for selected sites
}
