% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modules.R
\name{SimUI}
\alias{SimUI}
\alias{SimServer}
\title{Shiny Map UI for Projects}
\usage{
SimUI(
  id,
  label = "Simulation",
  custom_css_path = system.file("app/styles.css", package = "deersim"),
  custom_js_path = system.file("app/gomap.js", package = "deersim"),
  plm_labels = deersim::plm_labels
)

SimServer(id)
}
\arguments{
\item{id}{module id}

\item{label}{module label}

\item{custom_css_path}{custom css path for map}

\item{custom_js_path}{custom javascript path for module}

\item{plm_labels}{plm_labels}
}
\value{
shiny module

shiny module
}
\description{
Shiny module creating an interactive project map
}
\section{Functions}{
\itemize{
\item \code{SimServer()}: 

}}
